+++ CHANGES.rst
@@ -1,5 +1,18 @@
 CHANGELOG
 =========
+
+0.3.2
+-----
+- Added canned response.
+
+0.3.1
+-----
+- Order comments from newest to oldest unless they are replies.
+
+0.3
+---
+- Add canned responses
+
 0.2.9
 -----
 - Add support for replying to comments from the admin interface.